pyproject.toml
[project]
name = "codecracker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black",
    "langchain",
    "langchain-openai",
    "langchain-ollama",
    "bandit",
    "semgrep",
    "pandas",
    "python-dotenv",
    "tqdm",
    "python-dotenv",
    "datasets>=4.8.5",
    "jinja2>=3.1.6",
    "langchain-core>=1.3.2",
    "pylint>=4.0.5",
    "langsmith>=0.7.37",
]
